refactor: neuron_mean/norm
