+++ pytorch_optimizer/nero.py
@@ -3,26 +3,7 @@ from torch.optim.optimizer import Optimizer
 
 from pytorch_optimizer.base_optimizer import BaseOptimizer
 from pytorch_optimizer.types import CLOSURE, DEFAULTS, LOSS, PARAMETERS
-
-
-def neuron_norm(x: torch.Tensor):
-    if x.dim() <= 1:
-        return x.abs()
-
-    view_shape = [x.shape[0]] + [1] * (x.dim() - 1)
-    x = x.view(x.shape[0], -1)
-
-    return x.norm(dim=1).view(*view_shape)
-
-
-def neuron_mean(x: torch.Tensor) -> torch.Tensor:
-    if x.dim() <= 1:
-        raise ValueError('[-] neuron_mean not defined on 1D tensors.')
-
-    view_shape = [x.shape[0]] + [1] * (x.dim() - 1)
-    x = x.view(x.shape[0], -1)
-
-    return x.mean(dim=1).view(*view_shape)
+from pytorch_optimizer.utils import neuron_norm, neuron_mean
 
 
 class Nero(Optimizer, BaseOptimizer):
